+++ src/termclip.c
@@ -290,8 +290,8 @@ term_send_paste(void)
 void
 term_select_all(void)
 {
-  term.sel_start = (pos){-sblines(), 0, false};
-  term.sel_end = (pos){term_last_nonempty_line(), term.cols, true};
+  term.sel_start = (pos){-sblines(), 0, 0, 0, false};
+  term.sel_end = (pos){term_last_nonempty_line(), term.cols, 0, 0, true};
   term.selected = true;
   if (cfg.copy_on_select)
     term_copy();
@@ -313,29 +313,29 @@ term_get_text(bool all, bool screen, bool command)
 
     if (y < sbtop) {
       y = sbtop;
-      end = (pos){y, 0, false};
+      end = (pos){y, 0, 0, 0, false};
     }
     else {
       termline * line = fetch_line(y);
       if (line->lattr & LATTR_MARKED) {
         if (y > sbtop) {
           y--;
-          end = (pos){y, term.cols, false};
+          end = (pos){y, term.cols, 0, 0, false};
           termline * line = fetch_line(y);
           if (line->lattr & LATTR_MARKED)
             y++;
         }
         else {
-          end = (pos){y, 0, false};
+          end = (pos){y, 0, 0, 0, false};
         }
       }
       else {
         skipprompt = line->lattr & LATTR_UNMARKED;
-        end = (pos){y, term.cols, false};
+        end = (pos){y, term.cols, 0, 0, false};
       }
 
       if (fetch_line(y)->lattr & LATTR_UNMARKED)
-        end = (pos){y, 0, false};
+        end = (pos){y, 0, 0, 0, false};
     }
 
     int yok = y;
@@ -345,7 +345,7 @@ term_get_text(bool all, bool screen, bool command)
       printf("y %d skip %d marked %X\n", y, skipprompt, line->lattr & (LATTR_UNMARKED | LATTR_MARKED));
 #endif
       if (skipprompt && (line->lattr & LATTR_UNMARKED))
-        end = (pos){y, 0, false};
+        end = (pos){y, 0, 0, 0, false};
       else
         skipprompt = false;
       if (line->lattr & LATTR_MARKED) {
@@ -353,18 +353,18 @@ term_get_text(bool all, bool screen, bool command)
       }
       yok = y;
     }
-    start = (pos){yok, 0, false};
+    start = (pos){yok, 0, 0, 0, false};
 #ifdef debug_user_cmd_clip
     printf("%d:%d...%d:%d\n", start.y, start.x, end.y, end.x);
 #endif
   }
   else if (screen) {
-    start = (pos){term.disptop, 0, false};
-    end = (pos){term_last_nonempty_line(), term.cols, false};
+    start = (pos){term.disptop, 0, 0, 0, false};
+    end = (pos){term_last_nonempty_line(), term.cols, 0, 0, false};
   }
   else if (all) {
-    start = (pos){-sblines(), 0, false};
-    end = (pos){term_last_nonempty_line(), term.cols, false};
+    start = (pos){-sblines(), 0, 0, 0, false};
+    end = (pos){term_last_nonempty_line(), term.cols, 0, 0, false};
   }
   else if (!term.selected) {
     return wcsdup(W(""));
@@ -479,8 +479,8 @@ term_create_html(FILE * hf, int level)
   pos end = term.sel_end;
   bool rect = term.sel_rect;
   if (!term.selected) {
-    start = (pos){term.disptop, 0, false};
-    end = (pos){term.disptop + term.rows - 1, term.cols, false};
+    start = (pos){term.disptop, 0, 0, 0, false};
+    end = (pos){term.disptop + term.rows - 1, term.cols, 0, 0, false};
     rect = false;
   }
 
@@ -989,8 +989,8 @@ print_screen(void)
   else
     return;
 
-  pos start = (pos){term.disptop, 0, false};
-  pos end = (pos){term.disptop + term.rows - 1, term.cols, false};
+  pos start = (pos){term.disptop, 0, 0, 0, false};
+  pos end = (pos){term.disptop + term.rows - 1, term.cols, 0, 0, false};
   bool rect = false;
   clip_workbuf * buf = get_selection(start, end, rect, false);
   printer_wwrite(buf->text, buf->len);