+++ src/termout.c
@@ -2031,6 +2031,8 @@ set_modes(bool state)
           term.mouse_enc = state ? ME_UTF8 : 0;
         when 1006: /* Xterm's CSI-style mouse encoding */
           term.mouse_enc = state ? ME_XTERM_CSI : 0;
+        when 1016: /* Xterm's CSI-style mouse encoding with pixel resolution */
+          term.mouse_enc = state ? ME_PIXEL_CSI : 0;
         when 1015: /* Urxvt's CSI-style mouse encoding */
           term.mouse_enc = state ? ME_URXVT_CSI : 0;
         when 1037:
@@ -2232,6 +2234,8 @@ get_mode(bool privatemode, int arg)
         return 2 - (term.mouse_enc == ME_UTF8);
       when 1006: /* Xterm's CSI-style mouse encoding */
         return 2 - (term.mouse_enc == ME_XTERM_CSI);
+      when 1016: /* Xterm's CSI-style mouse encoding with pixel resolution */
+        return 2 - (term.mouse_enc == ME_PIXEL_CSI);
       when 1015: /* Urxvt's CSI-style mouse encoding */
         return 2 - (term.mouse_enc == ME_URXVT_CSI);
       when 1037: